+++ src/Sky.cpp
@@ -1,7 +1,7 @@
 #include "Sky.h"
 
 
-Sky::Sky(int width, int height): _width(width), _height(height)
+Sky::Sky(double width, double height): _width(width), _height(height)
 {
 
 }
@@ -11,5 +11,5 @@ Sky::~Sky()
 }
 
 // getter / setter
-int Sky::getWidth() { return _width; }
-int Sky::getHeight() { return _height; }
+double Sky::getWidth() { return _width; }
+double Sky::getHeight() { return _height; }
